Move some serialization code into middleware
