+++ src/frontend/serialization.rs
@@ -8,12 +8,9 @@ use crate::backends::plonky2::mock_main::MockMainPod;
 use crate::backends::plonky2::mock_signed::MockSignedPod;
 use crate::frontend::containers::Dictionary;
 use crate::frontend::Statement;
-use crate::middleware::{PodId, F};
-use crate::middleware::{HASH_SIZE, VALUE_SIZE};
-use plonky2::field::types::Field;
+use crate::middleware::PodId;
 
-use super::Value;
-use super::{MainPod, SignedPod};
+use super::{MainPod, SignedPod, Value};
 
 #[derive(Serialize, Deserialize, JsonSchema)]
 pub struct SignedPodHelper {
@@ -114,71 +111,6 @@ where
     String::deserialize(deserializer)?
         .parse()
         .map_err(serde::de::Error::custom)
-}
-
-fn serialize_field_tuple<S, const N: usize>(
-    value: &[F; N],
-    serializer: S,
-) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serializer.serialize_str(&format!(
-        "{:016x}{:016x}{:016x}{:016x}",
-        value[0].0, value[1].0, value[2].0, value[3].0
-    ))
-}
-
-fn deserialize_field_tuple<'de, D, const N: usize>(deserializer: D) -> Result<[F; N], D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let hex_str = String::deserialize(deserializer)?;
-
-    if !hex_str.chars().count() == 64 || !hex_str.chars().all(|c| c.is_ascii_hexdigit()) {
-        return Err(serde::de::Error::custom(
-            "Invalid hex string format - expected 64 hexadecimal characters",
-        ));
-    }
-
-    let mut v = [F::ZERO; N];
-    for i in 0..N {
-        let start = i * 16;
-        let end = start + 16;
-        let hex_part = &hex_str[start..end];
-        v[i] = F::from_canonical_u64(
-            u64::from_str_radix(hex_part, 16).map_err(serde::de::Error::custom)?,
-        );
-    }
-    Ok(v)
-}
-
-pub fn serialize_hash_tuple<S>(value: &[F; HASH_SIZE], serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serialize_field_tuple::<S, HASH_SIZE>(value, serializer)
-}
-
-pub fn deserialize_hash_tuple<'de, D>(deserializer: D) -> Result<[F; HASH_SIZE], D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    deserialize_field_tuple::<D, HASH_SIZE>(deserializer)
-}
-
-pub fn serialize_value_tuple<S>(value: &[F; VALUE_SIZE], serializer: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    serialize_field_tuple::<S, VALUE_SIZE>(value, serializer)
-}
-
-pub fn deserialize_value_tuple<'de, D>(deserializer: D) -> Result<[F; VALUE_SIZE], D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    deserialize_field_tuple::<D, VALUE_SIZE>(deserializer)
 }
 
 // HashMap is not ordered, but we want our dictionaries to be ordered
